Combine mentor and learner rows into a single row.
The intent of the separation was to allow distinct answers for
enrollment as a mentor and a learner, but in looking at the fields for
each, most are identical.

Note that this is a "lossy" migration (it deletes all rows).  That's OK
at this point because the application is not currently deployed.

This removes the enrollment form, as that was intended to connect to
Alchemer and is being replaced in #13.